V1.16.12 cybozu lbipam with 41147 and 41122 (#3)
* backport cilium/cilium/pull/41122

Signed-off-by: terashima <tomoya-terashima@cybozu.co.jp>

* backport cilium/cilium/pull/41147

lbipam: do not reallocate IPs on operator restart

If there was a pool that was filled and had unsatisfied Services, on
operator restart there was a high chance that we will reshuffle
assignement of IPs for that pool. This resulted in previously safisfied
services to either become unsatisfied or get a new IP.

Issue is fixed by not performing any operation on services until full
sync happens. After that, first we try to reuse IPs for already
satisfied services and only after that we try to assign additional IPs
to unsatisfied services.

Additionally, add test that covers this case, simulating restart of
operator.

Related: #40358

Signed-off-by: Marcel Zieba <marcel.zieba@isovalent.com>

---------

Signed-off-by: terashima <tomoya-terashima@cybozu.co.jp>
Signed-off-by: Marcel Zieba <marcel.zieba@isovalent.com>
Co-authored-by: Marcel Zieba <marcel.zieba@isovalent.com>